+++ src/aws/cis-1.4.0/rules/aws-cis-1.4.0-2.2.1.ts
@@ -54,4 +54,18 @@ export default {
   ],
 
   severity: 'medium',
+  gql: `{
+    queryawsEbs {
+      id
+      arn
+      accountId
+      __typename
+      encrypted
+    }
+  }`,
+  resource: 'queryawsEbs[*]',
+  conditions: {
+    path: '@.encrypted',
+    equal: true,
+  },
 }