+++ src/aws/cis-1.5.0/rules/index.ts
@@ -55,6 +55,7 @@ import Aws_CIS_150_412 from './aws-cis-1.5.0-4.12'
 import Aws_CIS_150_413 from './aws-cis-1.5.0-4.13'
 import Aws_CIS_150_414 from './aws-cis-1.5.0-4.14'
 import Aws_CIS_150_415 from './aws-cis-1.5.0-4.15'
+import Aws_CIS_150_416 from './aws-cis-1.5.0-4.16'
 import Aws_CIS_150_51 from './aws-cis-1.5.0-5.1'
 import Aws_CIS_150_52 from './aws-cis-1.5.0-5.2'
 import Aws_CIS_150_53 from './aws-cis-1.5.0-5.3'
@@ -119,6 +120,7 @@ export default [
   Aws_CIS_150_413,
   Aws_CIS_150_414,
   Aws_CIS_150_415,
+  Aws_CIS_150_416,
   Aws_CIS_150_51,
   Aws_CIS_150_52,
   Aws_CIS_150_53,